# Optimize AddPropertyIfAbsent