--- a/src/Serilog/Capturing/MessageTemplateProcessor.cs
+++ b/src/Serilog/Capturing/MessageTemplateProcessor.cs
@@ -14,7 +14,7 @@
 
 namespace Serilog.Capturing;
 
-class MessageTemplateProcessor : ILogEventPropertyFactory
+class MessageTemplateProcessor : ILogEventPropertyFactory, ILogEventPropertyValueFactory
 {
     readonly MessageTemplateCache _parser = new(new MessageTemplateParser());
     readonly PropertyBinder _propertyBinder;