--- a/test/Serilog.ApprovalTests/Serilog.approved.txt
+++ b/test/Serilog.ApprovalTests/Serilog.approved.txt
@@ -112,7 +112,7 @@ namespace Serilog.Core
     }
     public interface ILogEventPropertyValueFactory
     {
-        Serilog.Events.LogEventPropertyValue CreatePropertyValue(object value, bool destructureObjects = false);
+        Serilog.Events.LogEventPropertyValue CreatePropertyValue(object? value, bool destructureObjects = false);
     }
     public interface ILogEventSink
     {
@@ -629,7 +629,7 @@ namespace Serilog
         public static Serilog.ILogger Logger { get; set; }
         [Serilog.Core.MessageTemplateFormatMethod("messageTemplate")]
         public static bool BindMessageTemplate(string messageTemplate, object?[] propertyValues, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out Serilog.Events.MessageTemplate? parsedTemplate, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out System.Collections.Generic.IEnumerable<Serilog.Events.LogEventProperty>? boundProperties) { }
-        public static bool BindProperty(string propertyName, object value, bool destructureObjects, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out Serilog.Events.LogEventProperty? property) { }
+        public static bool BindProperty(string propertyName, object? value, bool destructureObjects, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out Serilog.Events.LogEventProperty? property) { }
         public static void CloseAndFlush() { }
         public static System.Threading.Tasks.ValueTask CloseAndFlushAsync() { }
         [Serilog.Core.MessageTemplateFormatMethod("messageTemplate")]